Increase the text size and add some plots Mina Ryten requested
I increased the text size by using "+ theme_grey(base_size = 18)" but
the code needs to be re-run to update the pdf files. I also increased the
width of the pdfs to 14 inches to give enough room for the larger
sub-tissue brain labels.

I added some code that plots the median and an interquartile range (IQR)-like
(2o to 80% because we don't have 25 and 75% quantile data saved) for the
expressed regions. For the exons, we would need to re-run `subtissue_exon.R`
to get these numbers.